- Add CMake code to detect and define fixed-size integral types in a   cross-platform-and-version-and-compiler way (argh!)
- Use CMake to detect header availability and use it, rather than
  platform macros like defined(_WIN32), to decide what headers to
  includes. Need to go through the sources and eliminate all use of
  defined(_WIN32) and the like in favour of function availablility
  tests, function compat wrappers, etc.

- Replace PDF_MIN and PDF_MAX macros with typesafe template versions
  that force the caller to be specific about the desired return type
  if there's a mismatch in the argument types. These versions also
  prevent the double-evaluation problem with the classic min/max macros.

- Begin moving platform-specific stuff to PdfCompilerCompat.{cpp,h}
  to keep the cruft away from the code that does the real work.

- Add podof_config.h generation to cmake, move many command line definitions
  into podofo_config.h for better visibility, more readable compiler command
  lines, etc.

- Add more common includes to pdfdefines

- Begin removing reundant includes from other source files

- Apply changes for Borland CC compatibility as desribed by Mirco Babin,
  with many changes/adjustments. Incomplete, requires testing with Borland
  and feedback.

- This provides the groundwork to start cleaning up the use of integer types
  in PoDoFo so it's clear what's used where and why, moving the platform-specific
  stuff into isolated wrappers, etc.



git-svn-id: https://svn.code.sf.net/p/podofo/code/podofo/trunk@1070 e6e4d1b7-ffb3-4f57-9d89-6d52b3892af5

+++ src/PdfFontCID.cpp
@@ -204,8 +204,8 @@ void PdfFontCID::CreateWidth( PdfObject* pFontDict ) const
         lGlyph = m_pMetrics->GetGlyphId( i );
         if( lGlyph )
         {
-            nMin = PDF_MIN( nMin, lGlyph );
-            nMax = PDF_MAX( nMax, lGlyph );
+            nMin = PDF_MIN( static_cast<long>(nMin), lGlyph );
+            nMax = PDF_MAX( static_cast<long>(nMax), lGlyph );
             nMax = PDF_MIN( nMax, cAbsoluteMax );
 
             if( lGlyph < cAbsoluteMax )
